Move Chat Screens to feature-chat root
